refactor(translator): share tool-arg helpers, cut double parse, forward late usage
Addresses Copilot review on PR #3779:
- Extract repairDuplicatedJsonArguments/appendToolArgs into
  translator/concerns/toolArgs.js shared by both Claude response
  translators (removes duplication drift risk).
- sanitizeToolArgs: single JSON.parse fast path; repair only on failure.
- Later finish chunk with usage now emits a usage-only message_delta
  (no duplicate tool blocks, no second message_stop) so clients still
  receive final usage.
