+++ open-sse/translator/response/openai-to-claude.js
@@ -3,6 +3,7 @@ import { FORMATS } from "../formats.js";
 import { ROLE, CLAUDE_BLOCK, MODEL_FALLBACK } from "../schema/index.js";
 import { fromOpenAIFinish } from "../concerns/finishReason.js";
 import { extractReasoningText } from "../concerns/reasoning.js";
+import { repairDuplicatedJsonArguments, appendToolArgs } from "../concerns/toolArgs.js";
 
 // Legacy "proxy_" prefix used by older request translators. Response strips it
 // defensively so tool names from such turns resolve back (e.g. proxy_Read → Read
@@ -10,89 +11,25 @@ import { extractReasoningText } from "../concerns/reasoning.js";
 // is then a no-op. Kept intentionally; do NOT couple to request's empty prefix.
 const CLAUDE_OAUTH_TOOL_PREFIX = "proxy_";
 
-// Sanitize tool call arguments to fix bad params from non-Anthropic models
+// Sanitize tool call arguments to fix bad params from non-Anthropic models.
+// Fast path: single parse for the common valid case; repair only on failure.
 function sanitizeToolArgs(toolName, argsJson) {
-  const repairedJson = repairDuplicatedJsonArguments(argsJson);
+  let args;
   try {
-    const args = JSON.parse(repairedJson);
-    const name = toolName.startsWith(CLAUDE_OAUTH_TOOL_PREFIX)
-      ? toolName.slice(CLAUDE_OAUTH_TOOL_PREFIX.length)
-      : toolName;
-    if (name === "Read") sanitizeReadArgs(args);
-    return JSON.stringify(args);
+    args = JSON.parse(argsJson);
   } catch {
-    return repairedJson;
-  }
-}
-
-// Repair duplicated / repeated JSON objects emitted by upstream proxies or cumulative streams
-function repairDuplicatedJsonArguments(raw) {
-  if (typeof raw !== "string" || raw.length < 4) return raw;
-  try {
-    JSON.parse(raw);
-    return raw;
-  } catch {}
-
-  const len = raw.length;
-  if (len % 2 === 0) {
-    const half = raw.slice(0, len / 2);
-    if (half === raw.slice(len / 2)) {
-      try {
-        JSON.parse(half);
-        return half;
-      } catch {}
+    const repairedJson = repairDuplicatedJsonArguments(argsJson);
+    try {
+      args = JSON.parse(repairedJson);
+    } catch {
+      return repairedJson;
     }
   }
-
-  const trimmed = raw.trim();
-  if (trimmed.startsWith("{")) {
-    let depth = 0;
-    let inString = false;
-    let escape = false;
-    for (let i = 0; i < trimmed.length; i++) {
-      const ch = trimmed[i];
-      if (escape) {
-        escape = false;
-        continue;
-      }
-      if (ch === "\\") {
-        escape = true;
-        continue;
-      }
-      if (ch === "\"") {
-        inString = !inString;
-        continue;
-      }
-      if (!inString) {
-        if (ch === "{") depth++;
-        else if (ch === "}") {
-          depth--;
-          if (depth === 0) {
-            const candidate = trimmed.slice(0, i + 1);
-            try {
-              JSON.parse(candidate);
-              const remainder = trimmed.slice(i + 1).trim();
-              if (remainder.startsWith("{")) {
-                return candidate;
-              }
-            } catch {
-              break;
-            }
-          }
-        }
-      }
-    }
-  }
-
-  return raw;
-}
-
-function appendToolArgs(current, incoming) {
-  if (!incoming) return current || "";
-  if (!current) return incoming;
-  if (incoming === current) return current;
-  if (incoming.startsWith(current)) return incoming;
-  return current + incoming;
+  const name = toolName.startsWith(CLAUDE_OAUTH_TOOL_PREFIX)
+    ? toolName.slice(CLAUDE_OAUTH_TOOL_PREFIX.length)
+    : toolName;
+  if (name === "Read") sanitizeReadArgs(args);
+  return JSON.stringify(args);
 }
 
 function sanitizeReadArgs(args) {
@@ -332,6 +269,16 @@ export function openaiToClaudeResponse(chunk, state) {
         usage: finalUsage
       });
       results.push({ type: "message_stop" });
+    } else if (state.usage && !state.messageStopSent) {
+      // Later finish chunk (commonly finish_reason:"stop" carrying usage).
+      // Tool blocks are already closed; emit a usage-only message_delta so the
+      // client still receives final usage without re-opening/duplicating blocks.
+      state.messageStopSent = true;
+      results.push({
+        type: "message_delta",
+        delta: { stop_reason: convertFinishReason(choice.finish_reason) },
+        usage: state.usage
+      });
     }
   }
